Refactor FastAPI sample with modular design
